move extract_endpoint files to proper place
